Disable 'model & hardware' dialog
The dialog is useless. It brings up the same information that you
already get from the preference's summary. I don't know why Google
decided to add this now. Maybe down the road they expand on it and
make it useful but I don't care for it in the current state.

Change-Id: I88f923a3697f20b89104a609410e051efe342ecc

Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>